requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
joblib==1.4.2
scikit-learn==1.5.0
tensorflow-cpu==2.16.1